requirements.txt
numpy
python-dotenv
opencv-contrib-python
mediapipe
protobuf==3.20